Revert "Check point"
This reverts commit 3eb6bbf65193044d2e5198ec85e2a58deadee0a6.

+++ CoAP.NET/ICoapConfig.cs
@@ -129,7 +129,7 @@ namespace Com.AugustCellars.CoAP
         /// <summary>
         /// Frequency to sweep for the sweep deduplicator
         /// </summary>
-        Int32 MarkAndSweepInterval { get; }
+        Int64 MarkAndSweepInterval { get; }
 
         /// <summary>
         /// Size of receive buffer if greater than 0
